Otimização dos tipos de dados inteiros.

+++ Paralelismo com Pthread/cristianolima_201700030145_pthread.c
@@ -7,27 +7,27 @@
 
 #define NUM_THREADS 4
 
-uint64_t quantidadeDeLinhasMatriz1;
-uint64_t quantidadeDeColunasMatriz1;
-uint64_t quantidadeDeLinhasMatriz2;
-uint64_t quantidadeDeColunasMatriz2;
+uint32_t quantidadeDeLinhasMatriz1;
+uint32_t quantidadeDeColunasMatriz1;
+uint32_t quantidadeDeLinhasMatriz2;
+uint32_t quantidadeDeColunasMatriz2;
 float *ptrMatriz1;
 float *ptrMatriz2;
 float *ptrMatrizResultante;
-uint64_t passo=0;
+uint32_t passo=0;
 
 void * multiplicaVetores(void *args){
-    uint64_t i = passo++;
+    uint32_t i = passo++;
 
-    for (int j = 0; j < quantidadeDeColunasMatriz2; j++)
-        for (int k = 0; k < quantidadeDeColunasMatriz1; k++){
+    for (uint32_t j = 0; j < quantidadeDeColunasMatriz2; j++)
+        for (uint32_t k = 0; k < quantidadeDeColunasMatriz1; k++){
             ptrMatrizResultante[i*quantidadeDeColunasMatriz2+j] += ptrMatriz1[ i*quantidadeDeColunasMatriz1+k ] * ptrMatriz2[ k*quantidadeDeColunasMatriz2+j ];
         }
     return 0;
 }
 
 int main(int argc, char *argv[]){
-    uint64_t quantidadeDeMultiplicacoes, quantidadeDeElementosMatriz1, quantidadeDeElementosMatriz2, quantidadeDeElementosMatrizResultante;
+    uint32_t quantidadeDeMultiplicacoes, quantidadeDeElementosMatriz1, quantidadeDeElementosMatriz2, quantidadeDeElementosMatrizResultante;
     uint32_t indexThread, indexDados, contadorFormatador;
     pthread_t threads[NUM_THREADS];
     bool fim=false;
@@ -35,11 +35,11 @@ int main(int argc, char *argv[]){
     FILE* fp_input = fopen( argv[1], "r" );
     FILE* fp_output = fopen( argv[2], "w" );
 
-    fscanf(fp_input, "%lu", &quantidadeDeMultiplicacoes);
+    fscanf(fp_input, "%u", &quantidadeDeMultiplicacoes);
     
     for(uint32_t contadorPrimeiroFor = 0; contadorPrimeiroFor < quantidadeDeMultiplicacoes; contadorPrimeiroFor++){
-        fscanf(fp_input, "%lu", &quantidadeDeLinhasMatriz1);
-        fscanf(fp_input, "%lu", &quantidadeDeColunasMatriz1);
+        fscanf(fp_input, "%u", &quantidadeDeLinhasMatriz1);
+        fscanf(fp_input, "%u", &quantidadeDeColunasMatriz1);
         
         quantidadeDeElementosMatriz1 = quantidadeDeLinhasMatriz1 * quantidadeDeColunasMatriz1;
         ptrMatriz1 = malloc(quantidadeDeElementosMatriz1 * sizeof(float));
@@ -48,8 +48,8 @@ int main(int argc, char *argv[]){
             fscanf(fp_input, "%f", &ptrMatriz1[contadorSegundoFor]);
         }
 
-        fscanf(fp_input, "%lu", &quantidadeDeLinhasMatriz2);
-        fscanf(fp_input, "%lu", &quantidadeDeColunasMatriz2);
+        fscanf(fp_input, "%u", &quantidadeDeLinhasMatriz2);
+        fscanf(fp_input, "%u", &quantidadeDeColunasMatriz2);
 
         quantidadeDeElementosMatriz2 = quantidadeDeLinhasMatriz2 * quantidadeDeColunasMatriz2;
         ptrMatriz2 = malloc(quantidadeDeElementosMatriz2 * sizeof(float));
